Moved api files out of app/lib/api into app/apis; renamed lib/api to lib/apiRequest; fixed index files and imports; fixed several instances of tokens being incorrectly typed. Note: temporarily left `lib/api/api.js.erb` at the same path, since apparently this path is hardcoded in the backend.
