+++ .flowconfig
@@ -11,6 +11,7 @@
 
 [options]
 # Allow importing directly from these folders.
+module.name_mapper='^apis' ->'<PROJECT_ROOT>/app/apis'
 module.name_mapper='^assets' ->'<PROJECT_ROOT>/app/assets'
 module.name_mapper='^config' ->'<PROJECT_ROOT>/app/config'
 module.name_mapper='^core-components' ->'<PROJECT_ROOT>/app/core-components'
