+++ app/apis/notifications/getAll.js
@@ -3,16 +3,16 @@
  * API docs: #TODO
  */
 
-import * as m from '../../model';
-import ApiRequest from '../../ApiRequest';
+import ApiRequest, { httpMethods, type ApiResponseData } from 'lib/ApiRequest';
+
 import { NOTIFICATIONS_ENDPOINT } from '../endpoints';
 
-const getAll = async (): Promise<m.ApiResponseData> => {
+const getAll = async (): Promise<ApiResponseData> => {
   const request = new ApiRequest();
 
   request
     .setEndpoint(NOTIFICATIONS_ENDPOINT)
-    .setMethod(m.httpMethods.GET)
+    .setMethod(httpMethods.GET)
     .setParameter('sort', '-createdAt')
     .setParameter('page[limit]', '10')
     .setParameter('page[offset]', '0')
