+++ app/apis/token/destroy.js
@@ -3,16 +3,16 @@
  * API docs: https://openwebslides.github.io/documentation/#token-api
  */
 
-import * as m from '../../model';
-import ApiRequest from '../../ApiRequest';
+import ApiRequest, { httpMethods, type ApiResponseData, type Token } from 'lib/ApiRequest';
+
 import { TOKEN_ENDPOINT } from '../endpoints';
 
-const destroy = (token: m.Token): Promise<m.ApiResponseData> => {
+const destroy = (token: Token): Promise<ApiResponseData> => {
   const request = new ApiRequest();
 
   request
     .setEndpoint(TOKEN_ENDPOINT)
-    .setMethod(m.httpMethods.DELETE)
+    .setMethod(httpMethods.DELETE)
     .setToken(token);
 
   return request.execute();
