+++ app/apis/topics/get.js
@@ -4,17 +4,16 @@
  */
 
 import type { Identifier } from 'types/model';
+import ApiRequest, { httpMethods, type ApiResponseData } from 'lib/ApiRequest';
 
-import * as m from '../../model';
-import ApiRequest from '../../ApiRequest';
 import { TOPICS_ENDPOINT } from '../endpoints';
 
-const get = async (id: Identifier): Promise<m.ApiResponseData> => {
+const get = async (id: Identifier): Promise<ApiResponseData> => {
   const request = new ApiRequest();
 
   request
     .setEndpoint(TOPICS_ENDPOINT)
-    .setMethod(m.httpMethods.GET)
+    .setMethod(httpMethods.GET)
     .setResource(id)
     .setParameter('include', 'user');
