+++ app/apis/topics/getAllByUserId.js
@@ -4,19 +4,18 @@
  */
 
 import type { Identifier } from 'types/model';
+import ApiRequest, { httpMethods, type ApiResponseData } from 'lib/ApiRequest';
 
-import * as m from '../../model';
-import ApiRequest from '../../ApiRequest';
 import { USERS_ENDPOINT, TOPICS_ENDPOINT } from '../endpoints';
 
-const getAllByUserId = async (userId: Identifier): Promise<m.ApiResponseData> => {
+const getAllByUserId = async (userId: Identifier): Promise<ApiResponseData> => {
   const request = new ApiRequest();
 
   request
     .setEndpoint(USERS_ENDPOINT)
     .setResource(userId)
     .setSubEndpoint(TOPICS_ENDPOINT)
-    .setMethod(m.httpMethods.GET);
+    .setMethod(httpMethods.GET);
 
   return request.execute();
 };
