+++ app/apis/topics/getContent.js
@@ -4,19 +4,18 @@
  */
 
 import type { Identifier } from 'types/model';
+import ApiRequest, { httpMethods, type ApiResponseData, type Token } from 'lib/ApiRequest';
 
-import * as m from '../../model';
-import ApiRequest from '../../ApiRequest';
 import { TOPICS_ENDPOINT, TOPICS_CONTENT_ENDPOINT } from '../endpoints';
 
-const getContent = (topicId: Identifier, token: string): Promise<m.ApiResponseData> => {
+const getContent = (topicId: Identifier, token: Token): Promise<ApiResponseData> => {
   const request = new ApiRequest();
 
   request
     .setEndpoint(TOPICS_ENDPOINT)
     .setResource(topicId)
     .setSubEndpoint(TOPICS_CONTENT_ENDPOINT)
-    .setMethod(m.httpMethods.GET)
+    .setMethod(httpMethods.GET)
     .setToken(token);
 
   return request.execute();
