+++ app/apis/users/get.js
@@ -4,17 +4,17 @@
  */
 
 import type { Identifier } from 'types/model';
+import ApiRequest, { httpMethods, type ApiResponseData, type Token } from 'lib/ApiRequest';
 
-import * as m from '../../model';
-import ApiRequest from '../../ApiRequest';
 import { USERS_ENDPOINT } from '../endpoints';
 
-const get = (id: Identifier, token: ?string): Promise<m.ApiResponseData> => {
+// #TODO why use ?Token here and Token everywhere else? @florian
+const get = (id: Identifier, token: ?Token): Promise<ApiResponseData> => {
   const request = new ApiRequest();
 
   request
     .setEndpoint(USERS_ENDPOINT)
-    .setMethod(m.httpMethods.GET)
+    .setMethod(httpMethods.GET)
     .setResource(id)
     .setToken(token);
