--- app/lib/api/index.js
@@ -1,12 +1,0 @@
-// @flow
-
-import type {
-  ApiResponseData,
-  Token,
-} from './model';
-
-export { default } from './apis';
-export type {
-  ApiResponseData,
-  Token,
-};
