+++ app/modules/authentication/actionTypes.js
@@ -1,6 +1,6 @@
 // @flow
 
-import type { Token } from 'lib/api';
+import type { Token } from 'lib/ApiRequest';
 import type { User } from 'modules/users';
 
 /* Action constants */
