+++ app/modules/authentication/saga/api/password.js
@@ -2,7 +2,7 @@
 
 import { call } from 'redux-saga/effects';
 
-import apis from 'lib/api';
+import apis from 'apis';
 
 import * as t from '../../actionTypes';
