+++ app/modules/topics/saga/api/patchContent.js
@@ -4,7 +4,7 @@ import { flashMessage, flashErrorMessage } from 'redux-flash';
 import { call, put, select } from 'redux-saga/effects';
 
 import authentication from 'modules/authentication';
-import apis from 'lib/api';
+import apis from 'apis';
 import apiRequestsStatus from 'modules/apiRequestsStatus';
 
 import * as t from '../../actionTypes';
