+++ app/modules/topics/saga/api/post.js
@@ -3,7 +3,7 @@
 import { call, select } from 'redux-saga/effects';
 
 import authentication from 'modules/authentication';
-import apis from 'lib/api';
+import apis from 'apis';
 
 import * as t from '../../actionTypes';
