+++ app/modules/users/saga/api/tests/users.test.js
@@ -2,7 +2,7 @@
 
 import { expectSaga } from 'redux-saga-test-plan';
 
-import apis from 'lib/api';
+import apis from 'apis';
 import authentication from 'modules/authentication';
 
 import * as t from '../../../actionTypes';
